% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trinomix.R
\name{sim_trinomix}
\alias{sim_trinomix}
\title{This is done with no stick breaking.}
\usage{
sim_trinomix(
  n_obs = 1000,
  n_groups = 10,
  ess_fraction = 1,
  tot_n = 100,
  p = NULL
)
}
\arguments{
\item{n_obs}{Number of observations (rows of data matrix to simulate). Defaults to 10}

\item{n_groups}{Number of categories for each observation (columns of data matrix). Defaults to 10}

\item{ess_fraction}{The effective sample size fraction, defaults to 1}

\item{tot_n}{The total sample size to simulate for each observation. This is approximate and the actual
simulated sample size will be slightly smaller. Defaults to 100}

\item{p}{The stock proportions to simulate from, as a vector. Optional, and when not included,
random draws from the dirichlet are used}
}
\description{
This is done with no stick breaking.
}
